feat-#24: move flight repository to db module
