+++ src/catalog/flight/commands/handlers/create-flight.handler.ts
@@ -3,7 +3,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 
 import { RpcExceptionService } from '../../../../utils/exception-handling';
 import { ErrorValidationService } from '../../../../utils/error-validation';
-import { FlightRepository } from '../../repositories';
+import { FlightRepository } from '../../../../db/repositories';
 import { CreateFlightCommand } from '../impl';
 import { FlightEntity } from '../../../../db/entities/flight.entity';
