+++ src/catalog/flight/commands/handlers/update-flight.handler.ts
@@ -2,7 +2,7 @@ import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 import { InjectRepository } from '@nestjs/typeorm';
 
 import { RpcExceptionService } from '../../../../utils/exception-handling';
-import { FlightRepository } from '../../repositories';
+import { FlightRepository } from '../../../../db/repositories';
 import { UpdateFlightCommand } from '../impl';
 import { FlightEntity } from '../../../../db/entities/flight.entity';
