+++ src/catalog/flight/flight.module.ts
@@ -6,7 +6,7 @@ import { UtilsModule } from '../../utils/utils.module';
 import { CommandHandlers } from './commands/handlers';
 import { FlightController } from './controllers/flight/flight.controller';
 import { QueryHandlers } from './queries/handlers';
-import { FlightRepository } from './repositories';
+import { FlightRepository } from '../../db/repositories';
 import { FlightService } from './services/flight/flight.service';
 
 @Module({
