+++ src/catalog/flight/queries/handlers/get-all-flights.handler.ts
@@ -2,7 +2,7 @@ import { IQueryHandler, QueryHandler } from '@nestjs/cqrs';
 import { InjectRepository } from '@nestjs/typeorm';
 
 import { RpcExceptionService } from '../../../../utils/exception-handling';
-import { FlightRepository } from '../../repositories';
+import { FlightRepository } from '../../../../db/repositories';
 import { GetAllFlightsQuery } from '../impl/get-all-flights.query';
 import { FlightEntity } from '../../../../db/entities/flight.entity';
