--- src/catalog/flight/repositories/index.ts
@@ -1,1 +1,0 @@
-export * from './flight.repository';
