+++ src/db/repositories/flight.repository.ts
@@ -1,6 +1,6 @@
 import { EntityRepository, Repository } from 'typeorm';
 
-import { FlightEntity } from '../../../db/entities/flight.entity';
+import { FlightEntity } from '../entities/flight.entity';
 
 @EntityRepository(FlightEntity)
 export class FlightRepository extends Repository<FlightEntity> {}
