+++ src/db/repositories/index.ts
@@ -1,2 +1,3 @@
 export * from './activity.repository';
 export * from './cars.repository';
+export * from './flight.repository';
